requirements.txt
newrelic==8.8.0
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
requests==2.28.2
urllib3==1.26.15
python-dotenv~=1.0.0